UserConfigPage 디자인 변경
