+++ app/src/main/java/com/wafflestudio/snutt2/views/RootActivity.kt
@@ -398,6 +398,7 @@ class RootActivity : AppCompatActivity() {
         composable2(NavigationDestination.TimeTableConfig) { TimetableConfigPage() }
         composable2(NavigationDestination.UserConfig) { UserConfigPage() }
         composable2(NavigationDestination.ChangeNickname) { ChangeNicknamePage() }
+        composable2(NavigationDestination.SocialLink) { SocialLinkPage() }
         composable2(NavigationDestination.PersonalInformationPolicy) { PersonalInformationPolicyPage() }
         composable2(NavigationDestination.ThemeModeSelect) { ColorModeSelectPage() }
         composable2(NavigationDestination.VacancyNotification) {
