+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/settings/SocialLinkViewModel.kt
@@ -1,0 +1,32 @@
+package com.wafflestudio.snutt2.views.logged_in.home.settings
+
+import androidx.lifecycle.ViewModel
+import com.wafflestudio.snutt2.data.user.UserRepository
+import com.wafflestudio.snutt2.lib.network.dto.GetUserFacebookResults
+import com.wafflestudio.snutt2.lib.network.dto.core.UserDto
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.StateFlow
+import javax.inject.Inject
+
+@HiltViewModel
+class SocialLinkViewModel @Inject constructor(
+    private val userRepository: UserRepository,
+) : ViewModel() {
+    val userInfo: StateFlow<UserDto?> = userRepository.user
+
+    suspend fun fetchUserInfo() {
+        userRepository.fetchUserInfo()
+    }
+
+    suspend fun fetchUserFacebook(): GetUserFacebookResults {
+        return userRepository.getUserFacebook()
+    }
+
+    suspend fun connectFacebook(id: String, token: String) {
+        userRepository.postUserFacebook(id, token)
+    }
+
+    suspend fun disconnectFacebook() {
+        userRepository.deleteUserFacebook()
+    }
+}
