+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/settings/UserConfigPage.kt
@@ -2,7 +2,6 @@ package com.wafflestudio.snutt2.views.logged_in.home.settings
 
 import android.os.Build
 import android.os.Build.VERSION_CODES
-import androidx.activity.result.ActivityResultRegistryOwner
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
@@ -26,11 +25,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.facebook.CallbackManager
-import com.facebook.FacebookCallback
-import com.facebook.FacebookException
-import com.facebook.login.LoginManager
-import com.facebook.login.LoginResult
 import com.wafflestudio.snutt2.R
 import com.wafflestudio.snutt2.components.compose.*
 import com.wafflestudio.snutt2.lib.android.toast
@@ -40,7 +34,6 @@ import com.wafflestudio.snutt2.ui.SNUTTTypography
 import com.wafflestudio.snutt2.views.*
 import com.wafflestudio.snutt2.views.logged_in.lecture_detail.Margin
 import kotlinx.coroutines.launch
-import timber.log.Timber
 
 @Composable
 fun UserConfigPage() {
@@ -56,51 +49,7 @@ fun UserConfigPage() {
 
     var addIdPasswordDialogState by remember { mutableStateOf(false) }
     var passwordChangeDialogState by remember { mutableStateOf(false) }
-    var disconnectFacebookDialogState by remember { mutableStateOf(false) }
     var leaveDialogState by remember { mutableStateOf(false) }
-
-    var facebookConnected by remember(user?.fbName) {
-        mutableStateOf(
-            user?.fbName.isNullOrEmpty().not(),
-        )
-    }
-
-    val callbackManager = CallbackManager.Factory.create()
-    LoginManager.getInstance()
-        .registerCallback(
-            callbackManager,
-            object : FacebookCallback<LoginResult> {
-                override fun onSuccess(result: LoginResult) {
-                    // App code
-                    val id = result.accessToken.userId
-                    val token = result.accessToken.token
-                    Timber.i("User ID: %s", result.accessToken.userId)
-                    Timber.i("Auth Token: %s", result.accessToken.token)
-                    // FIXME: staging 에서 테스트 불가
-                    scope.launch {
-                        launchSuspendApi(apiOnProgress, apiOnError) {
-                            viewModel.connectFacebook(id, token)
-                            facebookConnected = true
-                            viewModel.fetchUserFacebook().name
-                            // fetchUserInfo 와는 다른 성격? connectFacebook 이 성공하면
-                            // userInfo.fbName 항목은 따로 업데이트되지 않는건가?
-                        }
-                    }
-                }
-
-                override fun onCancel() {
-                    // App code
-                    Timber.w("Cancel")
-                    context.toast(context.getString(R.string.sign_up_facebook_login_failed_toast))
-                }
-
-                override fun onError(error: FacebookException) {
-                    // App code
-                    Timber.e(error)
-                    context.toast(context.getString(R.string.sign_up_facebook_login_failed_toast))
-                }
-            },
-        )
 
     Column(
         modifier = Modifier
@@ -172,35 +121,10 @@ fun UserConfigPage() {
                 }
             }
             Margin(height = 10.dp)
-            if (facebookConnected) {
-                SettingColumn {
-                    SettingItem(
-                        title = stringResource(R.string.settings_user_config_facebook_name),
-                        hasNextPage = false,
-                    ) {
-                        Text(
-                            text = user?.fbName ?: "",
-                            style = SNUTTTypography.body1.copy(
-                                color = SNUTTColors.Black500,
-                            ),
-                        )
-                    }
-                    SettingItem(
-                        title = stringResource(R.string.settings_user_config_facebook_disconnect),
-                        onClick = { disconnectFacebookDialogState = true },
-                    )
-                }
-            } else {
-                SettingItem(
-                    title = stringResource(R.string.settings_user_config_connect_facebook),
-                    onClick = {
-                        // FIXME: 실패했을 때.
-                        LoginManager.getInstance().logInWithReadPermissions(
-                            context as ActivityResultRegistryOwner, callbackManager, emptyList(),
-                        )
-                    },
-                )
-            }
+            SettingItem(
+                title = stringResource(R.string.social_link_title),
+                onClick = { navController.navigate(NavigationDestination.SocialLink) },
+            )
             Margin(height = 10.dp)
             SettingItem(
                 title = stringResource(R.string.settings_app_report_email),
@@ -352,25 +276,6 @@ fun UserConfigPage() {
             positiveButtonText = stringResource(R.string.settings_user_config_leave),
         ) {
             Text(text = stringResource(R.string.settings_leave_message), style = SNUTTTypography.body2)
-        }
-    }
-
-    if (disconnectFacebookDialogState) {
-        CustomDialog(
-            onDismiss = { disconnectFacebookDialogState = false },
-            onConfirm = {
-                scope.launch {
-                    launchSuspendApi(apiOnProgress, apiOnError) {
-                        viewModel.disconnectFacebook()
-                        LoginManager.getInstance().logOut()
-                        facebookConnected = false
-                        disconnectFacebookDialogState = false
-                    }
-                }
-            },
-            title = stringResource(R.string.settings_user_config_facebook_disconnect),
-        ) {
-            Text(text = stringResource(R.string.settings_user_config_disconnect_facebook_message), style = SNUTTTypography.body2)
         }
     }
 }
